rename: drop '_helpers' non-name; modules named by subject (#163)
idea_text_helpers → idea_naming (slug/tag/name operations)
idea_resonance_helpers → idea_resonance_tokens (token extraction)
idea_derive_helper → idea_derivation (synthetic-Idea construction)
idea_scoring_helpers → idea_scoring (cost/value/ROI math)

idea_internal_filter and idea_standing_questions kept (already named
by subject). Each module now answers 'what does this DO?' rather than
'what is this OTHER than the main thing?'

321 tests pass.
